requirements.txt
beautifulsoup4
customtkinter
python-dotenv
schedule
requests
lxml
feedparser